+++ test/login.test.js
@@ -53,9 +53,9 @@ describe('Test sidekick login', () => {
     const test = new SidekickTest({
       browser,
       page,
-      waitPopup: 4000,
-      waitNavigation: 'xx',
-      waitNavigationTime: 10000,
+      waitPopup: 2000,
+      // suppress extension hint
+      pre: (p) => p.evaluate(() => window.localStorage.setItem('hlxSidekickExtensionHint', Date.now() + 31536000000)),
       post: async (p) => {
         const btn = await p.waitForFunction(() => window.hlx.sidekick.shadowRoot.querySelector('.hlx-sk .user div.user-login button'));
         await btn.click();
@@ -117,6 +117,8 @@ describe('Test sidekick login', () => {
         status: 401,
       }],
       waitPopup: 2000,
+      // suppress extension hint
+      pre: (p) => p.evaluate(() => window.localStorage.setItem('hlxSidekickExtensionHint', Date.now() + 31536000000)),
       post: async (p) => {
         const btn = await p.waitForFunction(() => window.hlx.sidekick.shadowRoot.querySelector('.hlx-sk .user div.user-login button'));
         await btn.click();